Rename endpoint. Move related SDK and core to a new module.
