+++ api/test/integration/test_rbac_black_agent_endpoints.tavern.yaml
@@ -166,42 +166,6 @@ stages:
             time-reconnect: !anyint
 
 ---
-test_name: GET /agents/{agent_id}/cve
-
-stages:
-
-  - name: Get CVE information for agent 001
-    request:
-      verify: False
-      url: "{protocol:s}://{host:s}:{port:d}/agents/001/cve"
-      method: GET
-      headers:
-        Authorization: "Bearer {test_login_token}"
-      params:
-        limit: 1
-    response:
-      status_code: 200
-      json:
-        error: 0
-        data:
-          affected_items: !anything
-          failed_items: []
-          total_affected_items: !anyint
-          total_failed_items: 0
-
-  - name: Try to get CVE information for manager (000)
-    request:
-      verify: False
-      url: "{protocol:s}://{host:s}:{port:d}/agents/000/cve"
-      method: GET
-      headers:
-        Authorization: "Bearer {test_login_token}"
-      params:
-        limit: 1
-    response:
-      <<: *permission_denied
-
----
 test_name: GET /agents/{agent_id}/group/is_sync
 
 stages:
