+++ api/test/integration/test_rbac_white_all_endpoints.tavern.yaml
@@ -142,13 +142,13 @@ stages:
       <<: *permission_denied
 
 ---
-test_name: GET /agents/{agent_id}/cve
+test_name: GET /vulnerability/{agent_id}cve
 
 stages:
   - name: Try to get agent 000 vulnerabilities
     request:
       verify: False
-      url: "{protocol:s}://{host:s}:{port:d}/agents/000/cve"
+      url: "{protocol:s}://{host:s}:{port:d}/vulnerability/000"
       method: GET
       headers:
         Authorization: "Bearer {test_login_token}"
