+++ api/test/integration/test_rbac_white_vulnerability_endpoints.tavern.yaml
@@ -1,0 +1,41 @@
+
+---
+test_name: GET /vulnerability/{agent_id}
+
+marks:
+  - rbac_tests
+
+stages:
+
+  - name: Try to get vulnerabilities information for agent 001
+    request:
+      verify: False
+      url: "{protocol:s}://{host:s}:{port:d}/vulnerability/001"
+      method: GET
+      headers:
+        Authorization: "Bearer {test_login_token}"
+      params:
+        limit: 1
+    response: &permission_denied
+      status_code: 403
+      json:
+        error: 4000
+
+  - name: Get vulnerabilities information for manager (000)
+    request:
+      verify: False
+      url: "{protocol:s}://{host:s}:{port:d}/vulnerability/000"
+      method: GET
+      headers:
+        Authorization: "Bearer {test_login_token}"
+      params:
+        limit: 1
+    response:
+      status_code: 200
+      json:
+        error: 0
+        data:
+          affected_items: !anything
+          failed_items: []
+          total_affected_items: !anyint
+          total_failed_items: 0
