+++ framework/wazuh/rbac/default/relationships.yaml
@@ -28,6 +28,7 @@ relationships:
         - syscollector_read
         - logtest_all
         - task_status
+        - vulnerability_read
       rule_ids:
         - wui_elastic_admin
         - wui_opendistro_admin
@@ -45,6 +46,7 @@ relationships:
         - sca_read
         - syscheck_read
         - syscollector_read
+        - vulnerability_read
       rule_ids: []
 
     users_admin:
