+++ framework/wazuh/tests/data/security/rbac_catalog.yml
@@ -94,7 +94,6 @@ get_rbac_actions:
         related_endpoints:
           - GET /agents
           - GET /agents/{agent_id}/config/{component}/{configuration}
-          - GET /agents/{agent_id}/cve
           - GET /agents/{agent_id}/group/is_sync
           - GET /agents/{agent_id}/key
           - GET /agents/{agent_id}/stats/{component}
@@ -800,6 +799,20 @@ get_rbac_actions:
           effect: deny
         related_endpoints:
           - GET /tasks/status
+      vulnerability:read:
+        description: Allow read agents' vulnerabilities information
+        resources:
+          - agent:id
+          - agent:group
+        example:
+          actions:
+            - vulnerability:read
+          resources:
+            - agent:id:011
+            - agent:group:us-west
+          effect: allow
+        related_endpoints:
+          - GET /vulnerability/{agent_id}
 
 
 get_rbac_resources:
